pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-ai-bot"
version = "1.0.0"
description = "Modern async Telegram AI bot with PostgreSQL and Redis"
authors = [{name = "AI Assistant", email = "assistant@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "arq>=0.25.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "uvloop>=0.19.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
